requirements.txt
absl-py==1.4.0
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.5.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
black==23.11.0
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
clarabel==0.10.0
click==8.1.7
clip @ file:///mmfs1/gscratch/sewoong/anasery/rebasin_merging/transformer_models/CLIP
comm==0.2.0
contourpy==1.2.0
cvxpy==1.6.0
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
dm-tree==0.1.8
docker-pycreds==0.4.0
docstring-to-markdown==0.13
einops==0.7.0
etils==1.5.2
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fonttools==4.45.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.2.0
fvcore==0.1.5.post20221221
gdown==5.0.0
gitdb==4.0.11
GitPython==3.1.41
googleapis-common-protos==1.61.0
gurobipy==10.0.3
h5py==3.12.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.4
importlib-resources==6.1.1
iopath==0.1.10
ipykernel==6.26.0
ipython==8.18.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.12.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
lapjv==1.3.24
lightning-utilities==0.10.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open-clip-torch==2.24.0
osqp==0.6.7.post3
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
portalocker==2.8.2
prometheus-client==0.18.0
promise==2.3
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.9.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qdldl==0.1.7.post5
qtconsole==5.5.0
QtPy==2.4.1
referencing==0.31.0
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
safetensors==0.4.1
scikit-learn==1.4.0
scipy==1.11.4
scs==3.2.7.post2
Send2Trash==1.8.2
sentencepiece==0.2.0
sentry-sdk==1.39.2
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tensorflow-datasets==4.9.3
tensorflow-metadata==1.14.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.2.0
timm==0.9.12
tinycss2==1.2.1
tokenize-rt==5.2.0
toml==0.10.2
torch==2.1.2
torchmetrics==1.2.0
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
urllib3==2.1.0
wandb==0.16.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.3
zipp==3.17.0
